Meet new figma spec and move font size option  to ellipsis
